Updating several scripts

+++ site/profile/manifests/dm/dm_header_service.pp
@@ -10,15 +10,36 @@ class profile::dm::dm_header_service{
 		ts_sal_path => $ts_sal_path,
 	}
 
+	package { 'numpy':
+		ensure => installed,
+	}
+
+	package { 'PyYAML':
+		ensure => installed,
+	}
+
+	# EPEL is required to install astropy afterwards
+	package{ 'epel-release':
+		ensure => installed,
+		source => "http://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm",
+	}
+
+	# Required by the Header service
+	package{ 'python-astropy':
+		ensure => installed,
+		require => Package['epel-release']
+	}
+
 	exec { 'get-custom-fitsio':
 		command => 'wget https://github.com/menanteau/fitsio/archive/master.tar.gz -O /tmp/fitsio-master.tar.gz && cd /tmp/',
 		path => '/bin/',
+		onlyif => "test ! -f /tmp/fitsio-master.tar.gz"
 	}
 	
 	exec{ 'install-custom-fitsio':
 		path => '/bin/:/usr/local/bin',
 		cwd => "/tmp/",
-		command => "tar xvfz fitsio-master.tar.gz && cd fitsio-master && python3 setup.py install --prefix=/usr",
+		command => "tar xvfz fitsio-master.tar.gz && cd fitsio-master && python setup.py install --prefix=/usr",
 		onlyif => "test -f /tmp/fitsio-master.tar.gz"
 	}
 }
